Create CrudConfig and refactor project structure.
