+++ backend/context/build.gradle.kts
@@ -5,6 +5,4 @@ plugins {
 dependencies {
     implementation(kotlin("stdlib"))
     implementation(project(":backend:models"))
-    implementation(project(":backend:context"))
-    implementation(project(":backend:logics"))
 }
