+++ backend/context/src/main/kotlin/ru/otus/opinion/context/Context.kt
@@ -1,6 +1,7 @@
 package ru.otus.opinion.context
 
 import ru.otus.opinion.models.ServerError
+import ru.otus.opinion.models.State
 
 interface Context {
     var state: State
