+++ backend/logics/src/main/kotlin/ru/otus/opinion/logics/Crud.kt
@@ -4,7 +4,7 @@ import ru.otus.opinion.context.RequestContext
 import ru.otus.opinion.logics.chains.QuestionCreator
 import ru.otus.opinion.logics.chains.QuestionsListReader
 
-class Crud {
+class Crud(val config: CrudConfig = CrudConfig()) {
     suspend fun create(ctx: RequestContext) {
         QuestionCreator.process(ctx)
     }
