+++ backend/logics/src/main/kotlin/ru/otus/opinion/logics/CrudConfig.kt
@@ -1,0 +1,8 @@
+package ru.otus.opinion.logics
+
+import ru.otus.opinion.repo.api.Repo
+
+class CrudConfig (
+    val prodRepo: Repo = Repo.None,
+    val testRepo: Repo = Repo.None
+)
