+++ backend/logics/src/main/kotlin/ru/otus/opinion/logics/validation/CreateRequestValidator.kt
@@ -1,7 +1,7 @@
 package ru.otus.opinion.logics.validation
 
 import ru.otus.opinion.context.RequestContext
-import ru.otus.opinion.context.State
+import ru.otus.opinion.models.State
 import ru.otus.opinion.cor.dsl.builders.ChainBuilder
 import ru.otus.opinion.validation.validators.StringIsNotEmptyValidator
