+++ backend/logics/src/main/kotlin/ru/otus/opinion/logics/validation/dsl.kt
@@ -1,7 +1,7 @@
 package ru.otus.opinion.logics.validation
 
 import ru.otus.opinion.context.Context
-import ru.otus.opinion.context.State
+import ru.otus.opinion.models.State
 import ru.otus.opinion.cor.dsl.builders.ChainBuilder
 import ru.otus.opinion.models.ErrorLevel
 import ru.otus.opinion.models.ErrorType
