+++ backend/models/src/main/kotlin/ru/otus/opinion/models/ProcessingMode.kt
@@ -1,4 +1,4 @@
-package ru.otus.opinion.context
+package ru.otus.opinion.models
 
 enum class ProcessingMode {
     PROD, STUB, TEST
