+++ backend/models/src/main/kotlin/ru/otus/opinion/models/Stub.kt
@@ -1,4 +1,4 @@
-package ru.otus.opinion.context
+package ru.otus.opinion.models
 
 enum class Stub {
     NONE, SUCCESS, FAIL
