+++ backend/transport/mapping/build.gradle.kts
@@ -6,6 +6,7 @@ dependencies {
     implementation(kotlin("stdlib"))
     implementation(project(":backend:transport:openapi"))
     implementation(project(":backend:models"))
+    implementation(project(":backend:context"))
 
     testImplementation(kotlin("test"))
 }
