+++ backend/transport/mapping/src/test/kotlin/ru/otus/opinion/transport/mapping/MappingTest.kt
@@ -2,7 +2,7 @@ package ru.otus.opinion.transport.mapping
 
 import org.junit.Assert.assertTrue
 import ru.otus.opinion.context.RequestContext
-import ru.otus.opinion.context.State
+import ru.otus.opinion.models.State
 import ru.otus.opinion.models.*
 import ru.otus.opinion.models.stubs.QuestionStubs
 import ru.otus.opinion.openapi.transport.models.CreateQuestionResponse
